增加栏目文章映射关系

+++ internal/modules/Model.go
@@ -3,7 +3,7 @@ package modules
 import "time"
 
 type Model struct {
-	Id          uint    `gorm:"primary_key" json:"id"`
+	Id          uint64    `gorm:"primary_key" json:"id"`
 	CreatedBy   string    ` json:"createdBy"`
 	UpdateBy    string    `json:"updateBy"`
 	CreatedTime time.Time `json:"createdTime"`
